requirements.txt
flask
flask_cors
sqlalchemy
python-dotenv
psycopg2-binary
flask_sqlalchemy
flask_migrate
supabase